app/requirements.txt
feedparser==6.0.10
opml==0.5
Requests==2.31.0
starlette==0.31.1
multipart==0.2.4
bs4==0.0.1
SQLAlchemy==2.0.23
uvicorn==0.23.2
websockets==12.0
python-dateutil==2.8.2
Jinja2==3.1.2